Keep tasks declarations and definitions in the same file
