+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/task/LicenseReport.kt
@@ -30,9 +30,39 @@ import org.gradle.api.Task
 import org.gradle.api.tasks.TaskContainer
 
 /**
+ * Registers a [task][npmLicenseReport] for including
+ * NPM dependencies into a license report.
+ */
+fun JsTaskRegistering.licenseReport() {
+
+    // Temporarily, NPM dependencies are not included into license reports.
+    // See issue: https://github.com/SpineEventEngine/config/issues/301
+
+//    generateLicenseReport.finalizedBy(
+//        npmLicenseReport()
+//    )
+}
+
+
+/**
  * Locates `npmLicenseReport` task in this [TaskContainer].
  *
  * The task generates the report on NPM dependencies and their licenses.
  */
 internal val TaskContainer.npmLicenseReport: Task
     get() = getByName("npmLicenseReport")
+
+private fun JsTaskRegistering.npmLicenseReport() =
+    create("npmLicenseReport") {
+
+        description = "Generates the report on NPM dependencies and their licenses."
+        group = jsAnyTask
+
+        doLast {
+
+            // The script below generates license report for NPM dependencies and appends it
+            // to the report for Java dependencies generated by `generateLicenseReport` task.
+
+            npm("run", "license-report")
+        }
+    }
