--- buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/task/impl/LicenseReport.kt
@@ -1,58 +1,0 @@
-/*
- * Copyright 2021, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package io.spine.internal.gradle.javascript.task.impl
-
-import io.spine.internal.gradle.javascript.task.JsTaskRegistering
-
-/**
- * Registers a [task][io.spine.internal.gradle.javascript.task.npmLicenseReport] for including
- * NPM dependencies into a license report.
- */
-fun JsTaskRegistering.licenseReport() {
-
-    // Temporarily, NPM dependencies are not included into license reports.
-    // See issue: https://github.com/SpineEventEngine/config/issues/301
-
-//    generateLicenseReport.finalizedBy(
-//        npmLicenseReport()
-//    )
-}
-
-private fun JsTaskRegistering.npmLicenseReport() =
-    create("npmLicenseReport") {
-
-        description = "Generates the report on NPM dependencies and their licenses."
-        group = jsAnyTask
-
-        doLast {
-
-            // The script below generates license report for NPM dependencies and appends it
-            // to the report for Java dependencies generated by `generateLicenseReport` task.
-
-            npm("run", "license-report")
-        }
-    }
